Fixes + Trello Fixes
Menu Links have a broken black rectangle (focus behavior) on them when clicked to focus on that element.

Banner texts have bad practices for defining line heights which causes weird displays of text on them:

Content on the banner of the homepage is not centered (more to the right) because of min-width set to 110%

https://www.bradmontgomery.com/programs/online-motivational-speaker/

The second picture is not being shown fully it gets cut cut in half

https://www.bradmontgomery.com/contact/
The authenticator field’s styles are broken

https://www.bradmontgomery.com/

The font size/alignment is off. Also, the logo for Maryland SHRM is getting cut from the bottom.